+++ src/sentry/uuid.h
@@ -8,5 +8,6 @@ namespace sentry::uuid {
 
 String make_uuid();
 String make_uuid_no_dashes();
+String make_span_id();
 
 } // namespace sentry::uuid